Reorganization, cleanup items for wallet/publishers
