+++ js/wallet.js
@@ -7,30 +7,14 @@
 const Immutable = require('immutable')
 
 const Wallet = require('../abs/wallet')
-const {request} = require('../lib/util/request')
-const settings = require('../browser-laptop/js/constants/settings')
 const responses = require('../lib/data/responses.json')
 const ledgerStateData = require('../lib/data/ledger-state.json')
-
-const defaultAppState = Immutable.fromJS({
-  cache: {
-    ledgerVideos: {}
-  },
-  ledger: {
-    about: {}
-  },
-  about: {
-    preferences: {}
-  }
-})
 
 class JS extends Wallet {
   constructor () {
     super()
-    this.ledger = null
     this.wallet = 'default-wallet'
     this.stateKey = 'default-state'
-    this.state = defaultAppState
     this.stateFile = null
   }
 
@@ -51,16 +35,6 @@ class JS extends Wallet {
     this.setState(this.ledger.onInitRead(this.state, stateFile))
   }
 
-  get settingsP () {
-    return {
-      PAYMENTS_MINIMUM_VISITS: 1,
-      PAYMENTS_MINIMUM_VISIT_TIME: 8000,
-      PAYMENTS_CONTRIBUTION_AMOUNT: 10,
-      PAYMENTS_ENABLED: true,
-      PAYMENTS_NOTIFICATION_TRY_PAYMENTS_DISMISSED: true
-    }
-  }
-
   get cbResult () {
     const initialResult = Immutable.fromJS(JSON.parse(this.stateFile))
     const initialSeed = initialResult.getIn(['properties', 'wallet', 'keyinfo', 'seed'])
@@ -73,33 +47,8 @@ class JS extends Wallet {
     return result
   }
 
-  before (mockery) {
-    const self = this
-    mockery.enable({
-      warnOnReplace: false,
-      warnOnUnregistered: false,
-      useCleanCache: true
-    })
-
-    const fakeElectron = require('../test/fixtures/fakeElectron')
-    const fakeAdBlock = require('../test/fixtures/fakeAdBlock')
-    const fakeLevel = require('../test/fixtures/fakeLevel')
-
-    mockery.registerMock('electron', fakeElectron)
-    mockery.registerMock('level', fakeLevel)
-    mockery.registerMock('ad-block', fakeAdBlock)
-    mockery.registerMock('../../../js/lib/request', {
-      request: request
-    })
-    mockery.registerMock('../../../js/settings', {
-      getSetting: (key) => {
-        const keyP = Object.keys(settings).find((k) => settings[k] === key)
-        if (this.settingsP.hasOwnProperty(keyP)) {
-          return self.settingsP[keyP]
-        }
-      }
-    })
-
+  runBefore (mockery) {
+    this.before(mockery)
     this.setStateFile()
     this.ledger = require('../browser-laptop/app/browser/api/ledger')
     this.ledgerStatuses = require('../browser-laptop/app/common/constants/ledgerStatuses')
@@ -118,7 +67,7 @@ class JS extends Wallet {
   afterEach (mockery) {
     mockery.resetCache()
     this.setStateFile()
-    this.state = defaultAppState
+    this.state = this.defaultAppState
     this.ledger.resetModules()
   }
 
@@ -137,12 +86,12 @@ class JS extends Wallet {
   }
 
   enable () {
-    this.ledger.enable(defaultAppState)
+    this.ledger.enable(this.defaultAppState)
     return this.ledger.getSynopsis()
   }
 
   createWallet () {
-    this.setState(this.ledger.init(defaultAppState))
+    this.setState(this.ledger.init(this.defaultAppState))
     this.setState(this.ledger.onBootStateFile(this.state))
     this.setState(this.ledger.onCallback(this.state, this.cbResult, -1))
