+++ lib/helpers/stubs.js
@@ -139,8 +139,8 @@ module.exports = {
       name: 'setTestMinimums',
       func: function (self, state) {
         const synopsis = self.ledger.getSynopsis()
-        const minVisits = self.settingsP.PAYMENTS_MINIMUM_VISITS
-        const minDuration = self.settingsP.PAYMENTS_MINIMUM_VISIT_TIME
+        const minVisits = self.settings.PAYMENTS_MINIMUM_VISITS
+        const minDuration = self.settings.PAYMENTS_MINIMUM_VISIT_TIME
 
         synopsis.options.minPublisherVisits = minVisits
         synopsis.options.minPublisherDuration = minDuration
